Python Chart Code Version 5 and Text_Files Folder
This Python code entitled, "ChartNuclides_Dillmann_Version5.py" is a
program that displays the Chart of Nuclides based on user input. It will
show known experimental and theoretical information for each nuclide
pertaining to beta-delayed neutron emission, important for the r-process
which is responsible for the production of elements heavier than iron in
the Universe.

Note that the Text_Files folder in this repository contains all of the
relevant text files necessary for this program to run. A user must
obtain these text files and update the new location of these files in
the Python code before running.
Ex. filename1 = "C:\Users\Text_Files\N_Z_stable.txt"

This Chart code and this repository will be updated with more efficient
versions in the future.