BDNE Visualization Version 1
This update brings Version 1 of the BDNE Visualization Software which
includes a GUI for users to interact with.

More comments were included in Version 1. The GUI was finalized and the
code was reviewed and fixed for any immediate logic errors.

The Text Files folder was cleaned up as well as the addition of a folder
for older versions of this software.

Further improvements will continue as well as looking into introducing
different working versions of this software (in Python 3.3, iPython,
etc.).