BDNE Visualization Version 1.1 - Update 2
The 1.1 Version Update fixes a slight bug for users running the program
in Linux.

It also changes information found in the GUI slightly and it changes the
way P(xn) values are displayed in a visualization.

The Data File for the P(xn) experimental values has also been altered
for convenience.

BDNE Visualization Version 1 can be found in the folder 'Old
Visualization Versions'.